% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consensusSeekeR.R
\docType{data}
\name{A549_NR3C1_CFS_Peaks_partial}
\alias{A549_NR3C1_CFS_Peaks_partial}
\title{Sites with the greatest evidence of transcription factor binding
for the NR3C1 transcription factor from ENCODE (DDC accession: ENCFF002CFS).
For demonstration purpose.}
\format{A \code{GRanges} containing one entry per site. The peaks are
surronded by ranges present in the dataset
\code{A549_NR3C1_CFS_NarrowPeaks_partial}.}
\usage{
data(A549_NR3C1_CFS_Peaks_partial)
}
\description{
Sites representing the greatest evidence of enrichment for
the NR3C1 transcription factor (DCC accession: ENCFF002CFS)
for regions chr2:40000000-50000000 and chr3:10000000-13000000
from
the Encyclopedia of DNA Elements (ENCODE) data (Dunham I et al. 2012).
}
\details{
The peaks and ranges have been obtained using an optimal IDR analysis
done on all replicates.
}
\references{
\itemize{
    \item Dunham I, Kundaje A, Aldred SF, et al. An integrated encyclopedia
of DNA elements in the human genome. Nature. 2012 Sep 6;489(7414):57-74.
}
}
\seealso{
\itemize{
    \item \code{\link{A549_NR3C1_CFS_NarrowPeaks_partial}} { the associate
                 genomic regions dataset.}
    \item \code{\link{findConsensusPeakRegions}} {for extracting regions
                 sharing the same features in more than one experiment.}
}
}
\keyword{datasets}

