% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consensusSeekeR.R
\docType{package}
\name{consensusSeekeR-package}
\alias{consensusSeekeR}
\alias{consensusSeekeR-package}
\title{consensusSeekeR: Detection of consensus peak regions inside a group of
experiments using narrowPeak files}
\description{
This package compares multiple narrowPeak files to extract common
consensus peak regions. The size of the analyzed region is adjustable
as well as the number of narrowPeak files in which a peak must be detected
in the potential region to mark the region as a consensus peak region.
}
\author{
Astrid Louise Deschenes,
 Fabien Claude Lamaze,
 Pascal Belleau and
 Arnaud Droit

Maintainer: Astrid Louise Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\keyword{package}

