% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findCommonFeaturesIntern.R
\name{findCommonFeaturesForOneChrom}
\alias{findCommonFeaturesForOneChrom}
\title{TODO}
\usage{
findCommonFeaturesForOneChrom(chrName, padding, minNbrExp, allPeaks,
  allNarrowPeaks)
}
\arguments{
\item{chrName}{an object of \code{class} "formula" which contains a symbolic
model formula.}

\item{padding}{a \code{data.frame} containing the variables in the model.}

\item{minNbrExp}{}

\item{allPeaks}{a \code{GRanges} indicating which column from the
\code{data} must be added to the formula. When \code{NULL}, no
new term is added. Default : \code{NULL}.}

\item{allNarrowPeaks}{a \code{GRanges}.}
}
\value{
an object of \code{class} "commonFeatures".
}
\description{
TODO
}
\author{
Astrid Louise Deschenes
}
\keyword{internal}

