% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findCommonFeaturesIntern.R
\name{findCommonFeaturesValidation}
\alias{findCommonFeaturesValidation}
\title{TODO}
\usage{
findCommonFeaturesValidation(narrowpeaksBEDFiles, chrList, padding, minNbrExp,
  nbrThreads)
}
\arguments{
\item{narrowpeaksBEDFiles}{a \code{vector}}

\item{chrList}{a \code{vector}}

\item{padding}{a \code{numeric}}

\item{minNbrExp}{a \code{numeric}.}

\item{nbrThreads}{a \code{numeric}.}
}
\value{
\code{0}
}
\description{
TODO
}
\author{
Astrid Louise Deschenes
}
\keyword{internal}

