% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findCommonFeaturesIntern.R
\name{findConsensusPeakRegionsValidation}
\alias{findConsensusPeakRegionsValidation}
\title{Parameter validation for the \code{\link{findConsensusPeakRegions}}
     function}
\usage{
findConsensusPeakRegionsValidation(narrowPeakFiles, chrList, extendingSize,
  includeAllPeakRegion, minNbrExp, nbrThreads)
}
\arguments{
\item{narrowPeakFiles}{a \code{vector} containing the narrowPeak files to
use for the regions selection.}

\item{chrList}{a \code{vector} containing the name of the chromosomes to
analyze or the name \code{"ALL"} which indicate that all
chromosomes must be analyzed. When \code{NULL}, no
new term is added. Default : \code{NULL}.}

\item{extendingSize}{a \code{numeric} value indicating the size of padding
at each side of the peaks median position to create the consensus
region. The minimum size of the consensu region will be equal to
twice the value of the \code{extendingSize} parameter. The size of
the \code{extendingSize} must be a positive integer. Default = 250.}

\item{includeAllPeakRegion}{a \code{logical} indicating if the region set by
the \code{extendingSize} parameter is extended to include all
region of the peak closest to the peaks median position for each
experiment.}

\item{minNbrExp}{a \code{numeric} indicating the minimum number of BED files
in which a peak must be present for a region to be retained. The
numeric must be a positive value inferior or equal to the number of
files present in the \code{narrowpeaksBEDFiles} parameter.
Default = 1.}

\item{nbrThreads}{a \code{numeric} indicating the number of threads to use
         in parallel.}
}
\value{
\code{0} indicating that all parameters validations have been
     successful.
}
\description{
Validation of all parameters needed by the public
     \code{\link{findConsensusPeakRegions}} function.
}
\author{
Astrid Louise Deschenes
}
\keyword{internal}

