% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consensusSeekeR.R
\docType{data}
\name{Hosa_A549_FOSL2_01_NarrowPeaks_partial}
\alias{Hosa_A549_FOSL2_01_NarrowPeaks_partial}
\title{TODO Data Set}
\format{A \code{GRanges}}
\description{
TODO
}
\keyword{datasets}

