% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findConsensusPeakRegionsIntern.R
\name{isInteger}
\alias{isInteger}
\title{Validate if a value is an integer}
\usage{
isInteger(value)
}
\arguments{
\item{value}{an object to validate.}
}
\value{
\code{TRUE} is the parameter is a integer; otherwise \code{FALSE}
}
\description{
Validate if the value passed to the function is an integer or
         not. To be considered as an integer, the value must have a length
         of 1. The type of value can be a \code{integer} or
         \code{numerical}. However, a \code{numerical} must have the same
         value once casted to a \code{integer}.  A \code{vector} of
         integers will returned \code{FALSE}/
}
\author{
Astrid Louise Deschenes
}
\keyword{internal}

