% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readNarrowPeakIntern.R
\name{readNarrowPeak}
\alias{readNarrowPeak}
\title{TODO}
\usage{
readNarrowPeak(file_path)
}
\arguments{
\item{file_path}{a \code{vector}}
}
\value{
\code{0}
}
\description{
TODO
}
\author{
Astrid Louise Deschenes
}
\keyword{internal}

