% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sharedBed.R
\docType{package}
\name{sharedBed-package}
\alias{sharedBed}
\alias{sharedBed-package}
\title{sharedBed: Detection of shared features inside a group of experiences using
BED file}
\description{
This package compares multiple BED files to extract common
features. The size of the analyzed region is adjustable as well as the
number of BED files in which the features are detected to be retained.
}
\author{
Astrid Louise Deschenes

Maintainer: Astrid Louise Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\keyword{package}

