% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consensusSeekeR.R
\docType{data}
\name{A549_FOXA1_01_Peaks_partial}
\alias{A549_FOXA1_01_Peaks_partial}
\title{TODO Data Set}
\format{A \code{GRanges}}
\usage{
data(A549_FOXA1_01_Peaks_partial)
}
\description{
TODO Data Set
}
\references{
\itemize{
 \item Dunham I, Kundaje A, Aldred SF, et al. An integrated encyclopedia of
 DNA elements in the human genome. Nature. 2012 Sep 6;489(7414):57-74.
 }
}
\keyword{datasets}

