% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusSeekeR.R
\docType{package}
\name{consensusSeekeR-package}
\alias{consensusSeekeR-package}
\alias{consensusSeekeR}
\title{consensusSeekeR: Detection of consensus peak regions inside a group of
experiments using narrowPeak files}
\description{
This package compares positions and ranges data from multiple experiments
to extract common consensus regions. The size of the analyzed region is
adjustable as well as the number of experiments in which a peak must be
detected to mark a potential region as a consensus peak region.
}
\seealso{
\itemize{
\item \code{\link{readNarrowPeakFile}} {for extracting regions and peaks
from a narrowPeak file.}
\item \code{\link{findConsensusPeakRegions}} { for extracting regions
sharing the same features in more than one experiment. }
}
}
\author{
Astrid Deschenes,
Fabien Claude Lamaze,
Pascal Belleau and
Arnaud Droit

Maintainer:
Astrid Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\keyword{package}
