% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findCommonFeatures.R
\name{findCommonFeatures}
\alias{findCommonFeatures}
\title{TODO}
\usage{
findCommonFeatures(narrowpeaksBEDFiles, chrList = "ALL", padding = 250,
  minNbrExp = 1, nbrThreads = 1)
}
\arguments{
\item{narrowpeaksBEDFiles}{a \code{vector} containing the BED files to
use for the regions selection.}

\item{chrList}{a \code{vector} containing the name of the chromosomes to
analyze or the name \code{"ALL"} which indicate that all
chromosomes must be analyzed. When \code{NULL}, no
new term is added. Default : \code{NULL}.}

\item{padding}{a \code{numeric}. Default = 250.}

\item{minNbrExp}{a \code{numeric} indicating the minimum number of BED files
in which a peak must be present for a region to be retained. The
numeric must be a positive value inferior or equal to the number of
files present in the \code{narrowpeaksBEDFiles} parameter.
Default = 1.}

\item{nbrThreads}{a \code{numeric} indicating the number of threads to use
         in parallel.}
}
\value{
an object of \code{class} "commonFeatures".
}
\description{
TODO
}
\author{
Astrid Louise Deschenes
}

