% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findConsensusPeakRegions.R
\name{findConsensusPeakRegions}
\alias{findConsensusPeakRegions}
\title{Extract regions sharing the same features in more than one experiment}
\usage{
findConsensusPeakRegions(narrowPeaks, peaks, chrInfo, extendingSize = 250,
  includeAllPeakRegion = TRUE, minNbrExp = 1, nbrThreads = 1)
}
\arguments{
\item{narrowPeaks}{a \code{vector} containing \code{GRanges} representing
called peaks of signal enrichment based on pooled, normalized data
for all experiments.}

\item{peaks}{a \code{vector} containing \code{GRanges} representing peaks.}

\item{chrInfo}{a \code{Seqinfo} containing the name and the length of the
chromosomes to analyze.}

\item{extendingSize}{a \code{numeric} value indicating the size of padding
at each side of the peaks median position to create the consensus
region. The minimum size of the consensu region will be equal to
twice the value of the \code{extendingSize} parameter. The size of
the \code{extendingSize} must be a positive integer. Default = 250.}

\item{includeAllPeakRegion}{a \code{logical} indicating if the region size,
which is set by the \code{extendingSize} parameter is extended to
include all region of the peak closest to the peaks median
position for each experiment. When two peaks are at equal distance
of the peaks median for one experiment, both peaks are used to
extend the final consensus region.}

\item{minNbrExp}{a \code{numeric} indicating the minimum number of BED files
in which a peak must be present for a region to be retained. The
numeric must be a positive integer inferior or equal to the number
of files present in the \code{narrowPeakFiles} parameter.
Default = 1.}

\item{nbrThreads}{a \code{numeric} indicating the number of threads to use
         in parallel. The \code{nbrThreads} must be a positive integer.
         Default = 1.}
}
\value{
an object of \code{class} "consensusRanges".
}
\description{
Find regions sharing the same features for a minimum number of
         experiments using called peaks of signal enrichment based on
         pooled, normalized data (mainly coming from narrowPeak files).
         The peaks and narrow peaks and used to identify
         the common regions. The minimum number of experiments that must a
         peak in a common regions for that region to be selected is
         specified by user, as well as the size of padding. Only the
         chromosomes specified by the user are treated. The function can be
         parallized by specifying a number of threads superior to 1.

         When the padding is small, the detected regions are smaller than
         the one that could be obtained by doing an overlap of the narrow
         regions. Even more, the parameter specifying the minimum number of
         experiments needed to retain a region add versatility to the
         function.

         The side of the padding can have a large effect on the detected
         regions. It is recommanded to test more than one size and to do
         some manual validation of the resulting regions before selecting
         the final padding size.
}
\author{
Astrid Louise Deschenes
}

