% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findConsensusPeakRegions.R
\name{findConsensusPeakRegions}
\alias{findConsensusPeakRegions}
\title{Extract regions sharing the same features in more than one experiment}
\usage{
findConsensusPeakRegions(narrowPeaks, peaks, chrInfo, extendingSize = 250,
  includeAllPeakRegion = FALSE, shrinkToFitPeakRegion = FALSE,
  minNbrExp = 1L, nbrThreads = 1L)
}
\arguments{
\item{narrowPeaks}{a \code{GRanges} containing
called peak regions of signal enrichment based on pooled, normalized data
for all analyzed experiments. All \code{GRanges} entries must
have a metadata field called "name" which identifies the region to
the called peak. All \code{GRanges} entries must also
have a row name which identifies the experiment of origin. Each
\code{peaks} entry must have an associated \code{narrowPeaks} entry.
A \code{GRanges} entry is associated to a \code{narrowPeaks} entry by
having a identical metadata "name" field and a identical row name.}

\item{peaks}{a \code{GRanges} containing called peaks of signal
enrichment based on pooled, normalized data
for all analyzed experiments. All \code{GRanges} entries must
have a metadata field called "name" which identifies the called
peak. All \code{GRanges} entries must
have a row name which identifies the experiment of origin. Each
\code{peaks} entry must have an associated \code{narrowPeaks} entry. A
\code{GRanges} entry is associated to a \code{narrowPeaks} entry by having
a identical metadata "name" field and a identical row name.}

\item{chrInfo}{a \code{Seqinfo} containing the name and the length of the
chromosomes to analyze. Only the chomosomes contained in this
\code{Seqinfo} will be analyzed.}

\item{extendingSize}{a \code{numeric} value indicating the size of padding
on both sides of the position of the peaks median to create the
consensus region. The minimum size of the consensus region is
equal to twice the value of the \code{extendingSize} parameter.
The size of the \code{extendingSize} must be a positive integer.
Default = 250.}

\item{includeAllPeakRegion}{a \code{logical} indicating if the region size,
which is set by the \code{extendingSize} parameter is extended to include
the entire narrow peak regions of all peaks included in the unextended
consensus region. The narrow peak regions of the peaks added because of the
extension are not considered for the extension. Default: \code{FALSE}.}

\item{shrinkToFitPeakRegion}{a \code{logical} indicating if the region size,
which is set by the \code{extendingSize} parameter is shrinked to
fit the narrow peak regions of the peaks when all those regions
are smaller than the consensus region. Default: \code{FALSE}.}

\item{minNbrExp}{a \code{numeric} or a \code{integer} indicating the
minimum number of experiments in which at least one peak must be present
for a potential consensus region. The numeric must be a positive integer
inferior or equal to the number of experiments present in the
\code{narrowPeaks} and \code{peaks} parameters. Default = 1.}

\item{nbrThreads}{a \code{numeric} or a \code{integer} indicating the
number of threads to use in parallel. The \code{nbrThreads} must be a
positive integer. Default = 1.}
}
\value{
an object of \code{class} "consensusRanges" containing :
     \itemize{
         \item \code{call} the matched call.
         \item \code{consensusRanges} a \code{GRanges} containing the
consensus regions.
     }
}
\description{
Find regions sharing the same features for a minimum number of
experiments using called peaks of signal enrichment based on
pooled, normalized data (mainly coming from narrowPeak files).
The peaks and narrow peaks are used to identify
the consensus regions. The minimum number of experiments that must
have at least on peak in a region so that it is retained as a
consensus region is specified by user, as well as the size of
mining regions. Only the chromosomes specified by the user are treated.
The function can be parallized by specifying a number of threads superior
to 1. However, Windows does not support multicore  evaluation.

When the padding is small, the detected regions are smaller than
the one that could be obtained by doing an overlap of the narrow
regions. Even more, the parameter specifying the minimum number of
experiments needed to retain a region add versatility to the
function.

Beware that the side of the padding can have a large effect on
the detected consensus
regions. It is recommanded to test more than one size and to do
some manual validation of the resulting consensus regions before
selecting the final padding size.
}
\author{
Astrid Louise Deschenes
}

