% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consensusSeekeR.R
\docType{data}
\name{Hosa_A549_FOSL2_01_Peaks}
\alias{Hosa_A549_FOSL2_01_Peaks}
\title{Sites with the greatest evidence of transcription factor binding for the
FOSL2 transcription factor (for demonstration purpose)}
\format{A \code{GRanges} containing a metadata field called "name"
     which identified the peak. Each \code{GRanges} entry has a
     different peak name.}
\usage{
data(Hosa_A549_FOSL2_01_Peaks)
}
\description{
Sites with the greatest evidence of transcription factor binding for the
FOSL2 transcription factor (DCC accession: ENCFF000MZT), calculated
using the MACS peak caller (Zhang et al., 2008),
for regions chr1:249120200-249250621 and chr10:1-370100
from
the Encyclopedia of DNA Elements (ENCODE) data (Dunham I et al. 2012).
}
\references{
\itemize{
 \item Dunham I, Kundaje A, Aldred SF, et al. An integrated encyclopedia of
 DNA elements in the human genome. Nature. 2012 Sep 6;489(7414):57-74.
 \item Zhang Y, Liu T, Meyer CA, Eeckhoute J, Johnson DS, Bernstein BE, et
 al. Model-based analysis of ChIP-Seq (MACS). Genome  Biol. 2008, 9(9):R137.
 }
}
\seealso{
\itemize{
   \item \code{\link{Hosa_A549_FOSL2_01_NarrowPeaks}} { the associate
                 narrow peaks dataset}
   \item \code{\link{findConsensusPeakRegions}} {for extracting regions
                 sharing the same features in more than one experiment.}
}
}
\keyword{datasets}

