% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sharedBed.R
\docType{data}
\name{Hosa_A549_FOSL2_01_Peaks}
\alias{Hosa_A549_FOSL2_01_Peaks}
\title{TODO Data Set}
\format{A \code{GRanges}}
\usage{
data(Hosa_A549_FOSL2_01_Peaks)
}
\description{
TODO
}
\keyword{datasets}

