% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consensusSeekeR.R
\docType{data}
\name{Hosa_A549_FOSL2_01_Peaks_partial}
\alias{Hosa_A549_FOSL2_01_Peaks_partial}
\title{TODO Data Set}
\format{A \code{GRanges}}
\usage{
data(Hosa_A549_FOSL2_01_Peaks_partial)
}
\description{
TODO
}
\references{
\itemize{
 \item Dunham I, Kundaje A, Aldred SF, et al. An integrated encyclopedia of DNA
 elements in the human genome. Nature. 2012 Sep 6;489(7414):57-74.
 \item Zhang Y, Liu T, Meyer CA, Eeckhoute J, Johnson DS, Bernstein BE, et al.
 Model-based analysis of ChIP-Seq (MACS). Genome  Biol. 2008, 9(9):R137.
 }
}
\keyword{datasets}

